requirements.txt
uvicorn
flask
requests
streamlit
streamlit-option-menu
streamlit-modal

openpyxl
pandas
numpy
faiss-gpu
sentence-transformers
scikit-learn
langchain
langchain_openai
langchain_community
langchain_experimental
matplotlib
seaborn
tabulate